Fix endless recursion on correction for "..."
If a function makes use of varargs, but does not specify a required array type,
we raise an error and then rebuild the type.  Previously, some debug code was
checked in where instead of setting the type, we set the element type, so in
some cases the compiler would recurse endlessly when attempting to print an
error message.